requirements.txt
# Bot do Telegram
python-telegram-bot==20.6
python-dotenv==1.0.0

# API Backend
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.3.0
httpx>=0.25.0

# Integração com Supabase
supabase==1.0.3
postgrest==0.10.8

# Processamento de Pagamentos
mercadopago==2.2.0

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0 